n11_case/assignment4/requirements.txt
Flask==2.0.2
flask-socketio==5.1.1
kafka-python==2.0.2
eventlet==0.31.0
Werkzeug==2.0.0
gevent==23.9.1
dnspython==1.16.0
